+++ CHANGELOG.md
@@ -5,6 +5,13 @@
 This change log includes the summarized changes.
 For the full changelog, please go to the [Git Hub History][history] instead.
 
+### 2023.08.13
+
+- Added receptionist variants for the `My Personal Apartment` mod.
+- Added receptionist template pool.
+- Added receptionist spawn eggs and translations.
+- Fixed Key Holder Block rotations.
+
 ### 2023.08.01
 
 - Released first beta version for more detailed live testing.